push nourhan branch

+++ ApplicationB/Services B/Category/ProductCategoryService.cs
@@ -78,7 +78,7 @@ namespace ApplicationB.Services_B.Category
                 // Delete existing categories for the product
                 foreach (var item in existingProductCategories)
                 {
-                    await _repository.DeleteAsync(item.ProductId, item.CategoryId);
+                    await _repository.DeleteAsync((int)item.ProductId, item.CategoryId);
                 }
 
                 // Add new categories
@@ -139,7 +139,7 @@ namespace ApplicationB.Services_B.Category
         public async Task<ResultView<List<ProductCategoryDto>>> GetProductsByCategoryIdAsync(int categoryId)
         {
             var productCategories = await _repository.GetByCategoryIdAsync(categoryId);
-           
+
             if (!productCategories.Any())
             {
                 return ResultView<List<ProductCategoryDto>>.Failure($"No products found for category with ID '{categoryId}'.");
@@ -185,9 +185,9 @@ namespace ApplicationB.Services_B.Category
             }
 
             var mappedSubCategories = mainCategoriesWithSubcategories
-                .SelectMany(mc => mc.SubCategories) 
-                .Distinct() 
-                .Select(subCategory => _mapper.Map<GetAllCategoriesDTO>(subCategory)) 
+                .SelectMany(mc => mc.SubCategories)
+                .Distinct()
+                .Select(subCategory => _mapper.Map<GetAllCategoriesDTO>(subCategory))
                 .ToList();
 
             return ResultView<IEnumerable<GetAllCategoriesDTO>>.Success(mappedSubCategories);
